#33 refactor: 마이카이빙 관련 파일 네이밍 변경
