+++ android/MyCarchiving/app/src/main/java/com/softeer/mycarchiving/ui/myarchive/MyArchivingNavigation.kt
@@ -1,0 +1,21 @@
+package com.softeer.mycarchiving.ui.myarchive
+
+import androidx.navigation.NavController
+import androidx.navigation.NavGraphBuilder
+import androidx.navigation.NavOptions
+import androidx.navigation.compose.composable
+import com.softeer.mycarchiving.navigation.MainDestination
+
+fun NavController.navigateToMyArchiving(navOptions: NavOptions? = null) {
+    navigate(MainDestination.MY_ARCHIVING.route, navOptions)
+}
+
+fun NavGraphBuilder.myArchivingScreen(
+    onBackClick: () -> Unit
+) {
+    composable(route = MainDestination.MY_ARCHIVING.route) {
+        MyArchivingRoute(
+            onBackClick = onBackClick
+        )
+    }
+}
