+++ android/MyCarchiving/app/src/main/java/com/softeer/mycarchiving/ui/myarchive/MyArchivingViewModel.kt
@@ -1,4 +1,4 @@
-package com.softeer.mycarchiving.ui.mycar
+package com.softeer.mycarchiving.ui.myarchive
 
 import androidx.lifecycle.ViewModel
 import dagger.hilt.android.lifecycle.HiltViewModel
@@ -6,5 +6,5 @@ import javax.inject.Inject
 
 
 @HiltViewModel
-class MyCarViewModel @Inject constructor() : ViewModel() {
+class MyArchivingViewModel @Inject constructor() : ViewModel() {
 }
